Improved code readability and UI
